refactored code. moved all components into their own files. finished defeat screen. players can now click a button to get a new word from the same catagory or enter in a new catagory for a new word.
